% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Daily Closing Prices of major Stock Market Indices}
\format{An object of class \code{data.frame} with 4417 rows and 3 columns.}
\usage{
data
}
\description{
The file \code{stockmarket.RData} contains the daily closing prices of the DAX30, Dow Jones Industrials and the Nikkei 225.
}
\keyword{datasets}
