% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generator}
\alias{generator}
\title{Generate a time series containing a Visual Charting Pattern.}
\usage{
generator(start = 0, dlength = 100, tot.spread = 100, presig = 0,
  postsig = 0, plength = 5, parts = c(0, 15, 25, 50, 75, 85, 100),
  sprd = c(0, 50, 25, 100, 25, 50, 0))
}
\arguments{
\item{start}{Starting value}

\item{dlength}{Integer. Length of the signal part of the time series}

\item{tot.spread}{Integer. Difference between the lowest and highest value of the time series}

\item{presig}{Integer. Length of the pre-signal part of the time series}

\item{postsig}{Integer. Length of the post-signal part of the time series}

\item{plength}{Integer. Number of extrema in the pattern}

\item{parts}{Vector of Integers. Vector must contain plength + 2 elements, the first element being 0 and the last 100.
defines how far the extrema lie apart from another (in percent)}

\item{sprd}{Vector of Integers.Vector must contain plength + 2 elements, the first and last elemnt should be 0.
Defines the value of the extrema in percent of \code{tot.spread} in relation to \code{start}}
}
\value{
Time series with (optional) pre- or post signal, and the specified pattern.
}
\description{
Generate a time series containing a Visual Charting Pattern.
}
\examples{
#create a standard HS pattern:
generator()
#create a shifted head and shoulders pattern
generator(sprd = c(0,20,10,90,40,60,0))
#create a single peak, 10 data points, max is approximately 10
generator(0,10,10,0,0,1,c(0,50,100),c(0,100,0))

}

