% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter.R
\name{interpret}
\alias{interpret}
\title{Recognise patterns in Time Series Data}
\usage{
interpret(window, useriq = FALSE)
}
\arguments{
\item{window}{Time Series Data}

\item{useriq}{User-built recognition function. Set to \code{FALSE} if using inbuilt recognition capabilities.
This function must take a vector of 0s and 1s as first input and a vector of the extremum values as second input.
it should return the desired result.}
}
\description{
Recognise patterns in Time Series Data
}

