% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter.R
\name{iq}
\alias{iq}
\title{Inbuilt Recoqnition for 10 different financial markets patterns}
\usage{
iq(ext, exvals, expos, hsiq = TRUE, btpiq = TRUE, rtpiq = TRUE,
  dtpiq = TRUE)
}
\arguments{
\item{ext}{Extrema}

\item{exvals}{Values of the extrema}

\item{expos}{Position of the extrema}

\item{hsiq}{Logical. Recognise (inverse) Head and Shoulders pattern}

\item{btpiq}{Logical. Recognise Triangle and/or Broadening tops and bottoms pattern}

\item{rtpiq}{Logical. Recognise Rectangle tops and bottoms pattern}

\item{dtpiq}{Logical. Recognise Double tops and bottoms pattern}
}
\description{
Do not call individually. Switch recognition on/off for certain patterns, use the arguments in the \code{interpret} function.
}
