% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_additional.R
\name{loess.rpatrec}
\alias{loess.rpatrec}
\title{Uses the inbuilt function \code{loess()} to smooth the data using local polynomial regression.}
\usage{
loess.rpatrec(input, span = 0.75, ...)
}
\arguments{
\item{input}{Time series data passed to \code{loess()}}

\item{span}{The main smoothing parameter.}

\item{...}{Optional: Other arguments passed to \code{loess()}}
}
\value{
Smoothed time series data only, no additional output. Use \code{smooth.spline()} for greater functionality
}
\description{
This function is purely included to provide the standard interface coherent with other smoothers to the user.
}
