% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainput.R
\name{sample.pre}
\alias{sample.pre}
\title{Prepare Data for using with the other functions in the package}
\usage{
sample.pre(input)
}
\arguments{
\item{input}{Vector with time series data}
}
\value{
Returns a vector conataining time series data ready for further use.
}
\description{
This function removes bank holidays and other days for which only \code{NA} or repeated data is available.
Designed to work with Datastream data.
}
\details{
For an overview of the package capabilities, click here \link{rpatrec}.
}
