% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter.R
\name{slicer}
\alias{slicer}
\title{Break time series data into smaller "windows" and pass them to the interpret() function}
\usage{
slicer(data, length, step = 1, useriq = FALSE)
}
\arguments{
\item{data}{Time series Data}

\item{length}{Length of the "windows"}

\item{step}{Number of Data Points between windows}

\item{useriq}{User-built recognition function. Set to \code{FALSE} if using inbuilt recognition capabilities.
This function must take a vector of 0s and 1s as first input and a vector of the extremum values as second input.
It should return the desired result.}
}
\description{
Break time series data into smaller "windows" and pass them to the interpret() function
}

