% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stockmarket}
\alias{stockmarket}
\title{Daily Closing Prices of major Stock Market Indices}
\format{An object of class \code{data.frame} with 14852 rows and 6 columns.}
\usage{
stockmarket
}
\description{
A dataset containing the daily closing prices of the S&P500, DJ Industrials, the Milan Commit Index, etc.
}
\keyword{datasets}
