% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testfunction.R
\name{test.smoother}
\alias{test.smoother}
\title{Test how well a smoother can filter noise from data}
\usage{
test.smoother(n = 1, m = 5, incr = 1, max = 20, smoother,
  pattern = TRUE, ...)
}
\arguments{
\item{n}{number of runs}

\item{m}{number of runs per level of noise}

\item{incr}{value by which the error is increased in each turn}

\item{max}{max number of times the error is increased}

\item{smoother}{Function with pre-defined inputs, so that only the parameter \code{input} is left to be defined}

\item{pattern}{Check whether pattern was recognised. If \code{FALSE} only the correct position of extrema is chacked.}

\item{...}{other parameters the smoother requires}
}
\description{
Test how well a smoother can filter noise from data
}
