% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother2.r
\name{kernel}
\alias{kernel}
\title{Perform Kernel Regression on Time Series Data}
\usage{
kernel(input, bandwidth = "auto")
}
\arguments{
\item{input}{Vector of Time Series Data}

\item{bandwidth}{-Numerical: Choice of Bandwith
-\code{auto}: Choose bandwith by Cross validation automatically for the given sample}
}
\value{
Vector containing smoothed time series data, prints the bandwidth used.
}
\description{
Perform Kernel Regression on Time Series Data
}
