% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother1.R
\name{mav}
\alias{mav}
\title{Compute the moving average, exponential average or running median.}
\usage{
mav(input, len, method)
}
\arguments{
\item{input}{Time series data to use for computation}

\item{len}{An integer to determine the number of datapoints used for computation}

\item{method}{String. Determines the method of computation.
Permissible values are \code{exponential}, \code{simple} and \code{median}}
}
\value{
Vector containing computed values of length of \code{input} less length of \code{len}
}
\description{
Compute the moving average, exponential average or running median.
}
\examples{
mav(1:10,5,"exponential")

}

