% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{noise}
\alias{noise}
\title{Add noise to a time series}
\usage{
noise(input, type, level)
}
\arguments{
\item{input}{Time series to which noise will be added}

\item{type}{String. Defines what type of noise to be added. Acceptable values are:
\code{white} - white noise with variance \code{level}
\code{red} - red noise with variance \code{level} (use with caution)}

\item{level}{Number or \code{var}.
A number sets the standard deviation to a constant value at each point.
\code{var} sets the variance to increase proportional to the (absolute value of the) signal at peaks.}
}
\value{
Time series with added noise.
}
\description{
Add noise to a time series
}

