% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainput.R
\name{sample.pre}
\alias{sample.pre}
\title{Prepare Data for using with the other functions by removing non-numeric elements, and removing repeated values}
\usage{
sample.pre(input)
}
\description{
Prepare Data for using with the other functions by removing non-numeric elements, and removing repeated values
}

