% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother3.R
\name{savgolay}
\alias{savgolay}
\title{Perform Savitzgy-Golay smoothing on Time Series Data}
\usage{
savgolay(input, width = 4, degree = 2)
}
\arguments{
\item{input}{Input Vector with Time Series Data}

\item{width}{Width of the filter (to each side of the centre)}

\item{degree}{Highest degree polynomial}
}
\value{
Vector containing smoothed time series data.
}
\description{
Perform Savitzgy-Golay smoothing on Time Series Data
}
