% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_additional.R
\name{splines}
\alias{splines}
\title{Uses the inbuilt function \code{smooth.spline()} to smooth the data using splines.}
\usage{
splines(input, spar = 0.5, ...)
}
\arguments{
\item{input}{Time series data passed to \code{smooth.spline()}}

\item{spar}{Smoothing Parameter, value should be between 0 and 1. \code{NULL} for automatic computation}

\item{...}{Optional: Other arguments passed to \code{smooth.spline()}}
}
\value{
Smoothed time series data only, no additional output. Use \code{smooth.spline()} for greater functionality
}
\description{
This function is purely included to provide the standard interface coherent with other smoothers to the user.
}
